+++ test/ui-testing/license-terms.js
@@ -110,8 +110,8 @@ module.exports.test = (uiTestCtx) => {
 
       it('should open edit agreement page', done => {
         nightmare
-          .wait('#clickable-edit-agreement')
-          .click('#clickable-edit-agreement')
+          .wait('#clickable-edit-agreement') // edit button removed, ERM-693
+          .click('#clickable-edit-agreement') // edit button removed, ERM-693
           .wait('#formLicenses')
           .waitUntilNetworkIdle(1000)
           .then(done)