+++ test/ui-testing/orgs.js
@@ -200,8 +200,8 @@ module.exports.test = (uiTestCtx) => {
 
         it('should open edit Agreement', done => {
           nightmare
-            .wait('#clickable-edit-agreement')
-            .click('#clickable-edit-agreement')
+            .wait('#clickable-edit-agreement') // edit button removed, ERM-693
+            .click('#clickable-edit-agreement') // edit button removed, ERM-693
             .waitUntilNetworkIdle(2000)
             .then(done)
             .catch(done);