+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Modify keyboard shortcuts, use handlers from stripes-components, stripes-erm-components
 * Support order-lines 2.0 interface. ERM-1549
 * Added conditional rendering of accordions in Agreement view. ERM-1537
+* Removed edit button in agreements detail view. ERM-693
 ##  5.0.1 2020-11-05
 * Added permission check when displaying the add agreement line button. ERM-1197
 * Fixes bug where incorrect Dates are saved when tenant timezone is ahead of UTC. ERM-1202